1. 使用std::vector<std::vector<double*>> colMajorSparseMatrixRowId2DataPtr_来映射Sparse Hessian Matrix,提高了Construct_J_H_g的运行速度(首次需要使用三元组构建，耗时长)； 2. 同样使用Sparse H Matrix之后，计算理论残差下降值速度提高了10倍

+++ main_rgbd_slam.cpp
@@ -193,9 +193,13 @@ int main(int argc, char** argv) {
                 // 初始化深度图已经生成，后续需要对每一帧进行深度图传播
                 isInitialized = true;
                 cout << "\n******\nInitialized!\n******\n";
-                cv::imshow("init 2 KF", initer.debugMatchImg_);
-                cv::waitKey();
-                cv::destroyWindow("init 2 KF");
+                
+                if (config->debugShowOnlineResult3D) {
+                    cv::imshow("init 2 KF", initer.debugMatchImg_);
+                    cv::waitKey();
+                    cv::destroyWindow("init 2 KF");
+                }
+
                 // 模型预热
                 usleep(1000 * 1e3);
                 while (optimizer.newKF_ != nullptr) {
